ExperimentParameters: eps from float to double
Timestep type has been changed to double to avoid rounding issues
(visible in output filenames) with runs over 200 fm/c.

This commit refs #235

Reported-by: Dmitry Oliinychenko <oliiny@fias.uni-frankfurt.de>

+++ src/include/experimentparameters.h
@@ -18,7 +18,7 @@ struct ExperimentParameters {
   /// number of test particle
   int testparticles = 1;
   /// temporal time step
-  float eps = 0.001f;
+  double eps = 0.001;
   /// cross section of the elastic scattering
   float cross_section = 10.0f;
 };
